+++ EDDI/ChangeLog.md
@@ -11,6 +11,7 @@
 	* Added new 'SRV turret' event, triggered when you deploy or retract your SRV's turret
 	* Added new 'Ship fsd' event, triggered when there is a change to the status of your ship's fsd
 	* Added new 'Ship low fuel' event, triggered when your fuel level falls below 25%
+    * Added new 'Under attack' event
     * The 'Vehicle destroyed' event now includes the variable `vehicle`, describing the vehicle that was destroyed.
     * Add a new top level `status` object, which contains the following new variables
       * `vehicle` the vehicle that is under the commander's control.  Can be one of "Ship", "SRV" or "Fighter"